# Optimize roi_align on BMG