--- a/src/ATen/native/xpu/sycl/RoiAlignKernels.cpp
+++ b/src/ATen/native/xpu/sycl/RoiAlignKernels.cpp
@@ -69,23 +69,35 @@ T bilinear_interpolate(
 template <typename T>
 struct RoiAlignForwardKernel {
   void operator()(sycl::nd_item<1> item) const {
-    XPU_KERNEL_LOOP(item, index, nthreads_) {
-      // (n, c, ph, pw) is an element in the pooled output
+    auto wg = item.get_group(0);
+    auto idx = item.get_local_id(0);
+    int n = wg / wg_per_roi_;
+    int index =  (wg - n * wg_per_roi_) * item.get_local_range(0);
+    if (index < item_per_roi_) {
+
       int pw = index % pooled_width_;
       int ph = (index / pooled_width_) % pooled_height_;
       int c = (index / pooled_width_ / pooled_height_) % channels_;
-      int n = index / pooled_width_ / pooled_height_ / channels_;
 
       const T* offset_rois = rois_ + n * 5;
-      int roi_batch_ind = offset_rois[0];
+      if (idx==0) {
+      cache_roi_[0] = offset_rois[0];
 
       // Do not using rounding; this implementation detail is critical
       T offset = aligned_ ? (T)0.5 : (T)0.0;
-      T roi_start_w = offset_rois[1] * spatial_scale_ - offset;
-      T roi_start_h = offset_rois[2] * spatial_scale_ - offset;
-      T roi_end_w = offset_rois[3] * spatial_scale_ - offset;
-      T roi_end_h = offset_rois[4] * spatial_scale_ - offset;
-
+      cache_roi_[1] = offset_rois[1] * spatial_scale_ - offset;
+      cache_roi_[2] = offset_rois[2] * spatial_scale_ - offset;
+      cache_roi_[3] = offset_rois[3] * spatial_scale_ - offset;
+      cache_roi_[4] = offset_rois[4] * spatial_scale_ - offset;
+      }
+      item.barrier(sycl_local_fence);
+      
+      int roi_batch_ind = cache_roi_[0];
+      T roi_start_w = cache_roi_[1];
+      T roi_start_h = cache_roi_[2];
+      T roi_end_w = cache_roi_[3];
+      T roi_end_h = cache_roi_[4];
+      
       T roi_width = roi_end_w - roi_start_w;
       T roi_height = roi_end_h - roi_start_h;
       if (!aligned_) {
@@ -133,12 +145,15 @@ struct RoiAlignForwardKernel {
       output_val /= count;
 
       output_[index] = output_val;
+      
     }
   }
   RoiAlignForwardKernel(
       int nthreads,
       const T* input,
       const T spatial_scale,
+      int item_per_rois,
+      int wg_per_roi,
       int channels,
       int height,
       int width,
@@ -151,6 +166,8 @@ struct RoiAlignForwardKernel {
       : nthreads_(nthreads),
         input_(input),
         spatial_scale_(spatial_scale),
+	item_per_roi_(item_per_rois),
+	wg_per_roi_(wg_per_roi),
         channels_(channels),
         height_(height),
         width_(width),
@@ -160,20 +177,25 @@ struct RoiAlignForwardKernel {
         aligned_(aligned),
         rois_(rois),
         output_(output) {}
-
+  void sycl_ker_config_convention(sycl::handler& cgh) {
+    cache_roi_ = sycl_local_acc_t<T>(5, cgh);
+  }
  private:
-  int nthreads_;
+  const int nthreads_;
   const T* input_;
   const T spatial_scale_;
-  int channels_;
-  int height_;
-  int width_;
-  int pooled_height_;
-  int pooled_width_;
-  int sampling_ratio_;
-  bool aligned_;
+  const int item_per_roi_;
+  const int wg_per_roi_;
+  const int channels_;
+  const int height_;
+  const int width_;
+  const int pooled_height_;
+  const int pooled_width_;
+  const int sampling_ratio_;
+  const bool aligned_;
   const T* rois_;
   T* output_;
+  sycl_local_acc_t<T> cache_roi_;
 };
 
 template <typename T>
@@ -415,11 +437,7 @@ Tensor roi_align_kernel(
 
   at::Tensor output = at::zeros(
       {num_rois, channels, pooled_height, pooled_width}, input.options());
-
   auto output_size = num_rois * pooled_height * pooled_width * channels;
-  int64_t global_range =
-      ceil_div(static_cast<int64_t>(output_size), static_cast<int64_t>(512));
-  int64_t local_range = 512;
 
   if (output.numel() == 0) {
     return output;
@@ -433,10 +451,16 @@ Tensor roi_align_kernel(
       input.scalar_type(),
       "roi_align_forward_kernel_xpu",
       [&] {
+        int64_t local_range = syclMaxWorkGroupSize<RoiAlignForwardKernel<scalar_t>>();
+	int item_per_roi = pooled_height * pooled_width * channels;
+	int wg_per_roi = (item_per_roi + local_range - 1) / local_range;
+	int64_t global_range= wg_per_roi * num_rois;
         auto kfn = RoiAlignForwardKernel<scalar_t>(
             output_size,
             input_.data_ptr<scalar_t>(),
             spatial_scale,
+	    item_per_roi,
+	    wg_per_roi,
             channels,
             height,
             width,