For input [1, 2048, 50, 75], rois [1000,5], roi align takes 4.7 ms on PVC but 75 ms on BMG. Each roi will have 2048x50x75 work items reading the same value from LLC, and it's very slow on BMG. After opt, PVC takes 2.42ms, BMG reaches  5.7ms